fix(entity): ECS反向桥接EntityOwnerComponent接线+horse/Wolf/Spear/序列化测试修复
根因:EntityOwnerComponent设计为持unique_ptr<Entity>(拥有),与EntityManager::m_entities
的unique_ptr形成双重拥有致double-free,故该组件此前从未被任何地方attach。导致
FireTickSystem/PortalTickSystem的view<...,EntityOwnerComponent>永远为空——
两个系统在生产与测试中均从未工作过(fire永不递减/portal永不计时)。

修复:
1. EntityOwnerComponent: unique_ptr<Entity>→Entity*非拥有裸指针(对齐基岩版
   ActorOwnerComponent持Actor&非拥有)。所有权仍归EntityManager::m_entities或
   测试局部变量,本组件仅反查。
2. Entity构造末尾self-attach EntityOwnerComponent{this},生产+测试全覆盖。
3. Entity析构新增:经m_entityContext->registry().destroy()销毁entt实体及全部组件
   (含EntityOwnerComponent),消除悬垂反查指针。valid()校验防重复destroy。
   成员析构顺序保证~Entity body先执行(m_entityContext仍存活)。
4. LivingEntity::setHealth置m_healthSynced=true,避免tick()首帧兜底setHealth
   (maxHealth)覆盖测试/业务构造后手动设置的health(致死状态isDying/deathTime不更新)。
5. AbstractHorseEntity构造移除registerGoals调用(累加语义,子类构造再调致基础目标
   重复注册);5个子类(Horse/Donkey/Mule/SkeletonHorse/ZombieHorse)构造补调
   registerGoals。registerAttributes保持(setBaseValue幂等)。
6. testEcsRegistry helper: TestSerializationWorld补entityRegistry()重写返回
   &testEcsRegistry()(attachPassengers反序列化乘客须经world.entityRegistry()取
   ECS registry,否则MC_ASSERT_RELEASE(nullptr)崩)。
7. WolfEntityTest: DataParameter类型匹配修复(CollarColor注册i32读i32/
   AngerTime注册i64读i64,类型不符std::get抛bad variant access)。
8. AbstractHorseEntityTest/SpearItemTest: NBT序列化改writeToNBT/readFromNBT
   (addAdditionalSaveData已空壳,字段搬ComponentSerializerRegistry)+补
   VanillaEntities::registerAll()注册序列化器。Spear键名Item→item(小写,
   AbstractArrow子树)+移除knockback断言(运行时从附魔读不持久化)。
9. HorseAiGoalsTest: Llama目标数预期8→10(基类8+llama特定2)。
10. EntityLavaFireTest: baseTick后补ecs::FireTickSystem{}.tick(testEcsRegistry())
    驱动fire链(已迁出baseTick)。53用例全过。

验证: mc_tests EntityLavaFireTest 53/53 PASSED, LivingEntityTest.Death通过。

+++ tests/common/entity/entities/passive/tamable/WolfEntityTest.cpp
@@ -827,7 +827,7 @@ TEST_F(WolfEntityTestFixture, DataParameter_GetCollarColor_ReadsFromDataManager)
     auto& dataManager = wolf.dataManager();
     u16 paramId = WolfEntity::getCollarColorParamId();
     EXPECT_TRUE(dataManager.hasParam(paramId));
-    i32 storedValue = static_cast<i32>(dataManager.get<i64>(entity::DataParameter<i64>(paramId)));
+    i32 storedValue = static_cast<i32>(dataManager.get<i32>(entity::DataParameter<i32>(paramId)));
     EXPECT_EQ(storedValue, static_cast<i32>(DyeColor::Blue));
 }
 
@@ -839,7 +839,7 @@ TEST_F(WolfEntityTestFixture, DataParameter_SetCollarColor_WritesToDataManager)
 
     // 设置颈圈颜色
     wolf.setCollarColor(DyeColor::Green);
-    i32 storedValue = static_cast<i32>(dataManager.get<i64>(entity::DataParameter<i64>(paramId)));
+    i32 storedValue = static_cast<i32>(dataManager.get<i32>(entity::DataParameter<i32>(paramId)));
     EXPECT_EQ(storedValue, static_cast<i32>(DyeColor::Green));
 
     // 设置为白色
@@ -896,7 +896,7 @@ TEST_F(WolfEntityTestFixture, DataParameter_CollarColor_AllDyeColorsRoundTrip)
         DyeColor color = static_cast<DyeColor>(i);
         wolf.setCollarColor(color);
         EXPECT_EQ(wolf.getCollarColor(), color);
-        i32 storedValue = static_cast<i32>(dataManager.get<i64>(entity::DataParameter<i64>(paramId)));
+        i32 storedValue = static_cast<i32>(dataManager.get<i32>(entity::DataParameter<i32>(paramId)));
         EXPECT_EQ(storedValue, i);
     }
 }
@@ -3001,7 +3001,8 @@ TEST_F(WolfEntityTestFixture, DataParameter_SetAngry_WritesToDataManager)
     // 清除愤怒状态
     wolf.setAngry(false);
     EXPECT_FALSE(wolf.isAngry());
-    storedValue = dataManager.get<i32>(entity::DataParameter<i32>(paramId));
+    // DATA_ANGER_TIME_PARAM 注册为 i64（对齐 vanilla Long），须以 get<i64> 读取。
+    storedValue = static_cast<i32>(dataManager.get<i64>(entity::DataParameter<i64>(paramId)));
     EXPECT_EQ(storedValue, 0);
 }
 
@@ -3018,7 +3019,8 @@ TEST_F(WolfEntityTestFixture, DataParameter_SetAngerTime_WritesToDataManager)
 
     // 设置为 0
     wolf.setAngerTime(0);
-    storedValue = dataManager.get<i32>(entity::DataParameter<i32>(paramId));
+    // DATA_ANGER_TIME_PARAM 注册为 i64（对齐 vanilla Long），须以 get<i64> 读取。
+    storedValue = static_cast<i32>(dataManager.get<i64>(entity::DataParameter<i64>(paramId)));
     EXPECT_EQ(storedValue, 0);
 }
